[Chore] #66 - Domain 구현에 맞게 Push, Notification 폴더링 구분
