+++ Projects/Data/NotificationData/Sources/Factory/File.swift
@@ -1,0 +1,9 @@
+//
+//  File.swift
+//  NotificationData
+//
+//  Created by YunhakLee on 3/26/26.
+//  Copyright © 2026 kr.websoso.app. All rights reserved.
+//
+
+import Foundation
