+++ Projects/Data/NotificationData/Sources/Notification/Service/DefaultNotificationService.swift
@@ -10,15 +10,13 @@
 import Foundation
 import Networking
 
-final class DefaultNotificationService: NotificationService {
+struct DefaultNotificationService: NotificationService {
     private let client: NetworkingRequestable
     
     init(client: NetworkingRequestable) {
         self.client = client
     }
-    
-    // MARK: - 앱 내 알림
-    
+
     func getNotifications(_ query: NotificationQeury) async throws -> NotificationListResponse {
         let endpoint = NotificationEndpoint.getNotifications(query)
         return try await client.request(endpoint, decodeTo: NotificationListResponse.self)
@@ -37,22 +35,5 @@ final class DefaultNotificationService: NotificationService {
     func postNotificationRead(notificationId: Int) async throws {
         let endpoint = NotificationEndpoint.postNotificationRead(notificationId: notificationId)
         _ = try await client.request(endpoint)
-    }
-    
-    // MARK: - Push 알림
-    
-    func postFCMToken(_ request: FCMTokenRequest) async throws {
-        let endpoint = NotificationEndpoint.postFCMToken(request)
-        _ = try await client.request(endpoint)
-    }
-    
-    func postPushNotificationSetting(_ request: PushNotificationSettingRequest) async throws {
-        let endpoint = NotificationEndpoint.postPushNotificationSetting(request)
-        _ = try await client.request(endpoint)
-    }
-    
-    func getPushNotificationSetting() async throws -> PushNotificationSettingResponse {
-        let endpoint = NotificationEndpoint.getPushNotificationSetting
-        return try await client.request(endpoint, decodeTo: PushNotificationSettingResponse.self)
     }
 }
